More fixing warnings

+++ include/util/util.h
@@ -3,4 +3,4 @@
 #define FNV_OFFSET_BASIS 0xcbf29ce484222325
 #define FNV_PRIME 0x100000001b3
 
-unsigned long util_fnv1a_hash(const char* key);
+unsigned long long util_fnv1a_hash(const char* key);
